Move some of the selenium stuff into a base.
