+++ tests/sync/test_selenium_appeal_generation.py
@@ -8,16 +8,13 @@ import sys
 import pytest
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 from fighthealthinsurance.models import *
+from .fhi_selenium_base import FHISeleniumBase
 from seleniumbase import BaseCase
-from selenium.webdriver.common.by import By
-from selenium.webdriver.support.ui import WebDriverWait
-from selenium.webdriver.support import expected_conditions as EC
-
 
 BaseCase.main(__name__, __file__)
 
 
-class SeleniumTestAppealGeneration(BaseCase, StaticLiveServerTestCase):
+class SeleniumTestAppealGeneration(FHISeleniumBase, StaticLiveServerTestCase):
     fixtures = ["fighthealthinsurance/fixtures/initial.yaml"]
 
     @classmethod
@@ -29,38 +26,6 @@ class SeleniumTestAppealGeneration(BaseCase, StaticLiveServerTestCase):
     def tearDownClass(cls):
         super(StaticLiveServerTestCase, cls).tearDownClass()
         super(BaseCase, cls).tearDownClass()
-
-    def assert_title_eventually(self, desired_title):
-        try:
-            WebDriverWait(self.driver, 15).until(EC.title_is(desired_title))
-        except Exception:
-            # On failure assert_title gives us a better error message than the timeout.
-            self.assert_title(desired_title)
-
-    def assert_text_eventually(self, expected_text, target):
-        try:
-            element_locator = (By.ID, target)
-            WebDriverWait(self.driver, 15).until(
-                lambda d: d.find_element(*element_locator).text == expected_text
-            )
-        except Exception as e:
-            # On failure assert_text gives us a better error message than the timeout.
-            self.assert_text(
-                expected_text,
-                target)
-
-    def assert_text_eventually_contains(self, expected_text, target):
-        element_locator = (By.ID, target)
-        WebDriverWait(self.driver, 15).until(
-            lambda d: expected_text in d.find_element(*element_locator).text
-        )
-
-    def click_button_eventually(self, target):
-        element_locator = (By.ID, target)
-        WebDriverWait(self.driver, 60).until(
-            lambda d: d.find_element(*element_locator) != None
-        )
-        return self.click(f"button#{target}")
 
     def test_submit_an_appeal_with_missing_info_and_fail(self):
         self.open(f"{self.live_server_url}/")
